requirements.txt
flask
cryptography
psycopg2-binary
werkzeug